server/services/messages/requirements.txt
pymongo==3.9.0
emitter-io==2.0.3
